Metadata changes, make executor registrable
