+++ docs/source/api/components/step_info.rst
@@ -6,3 +6,21 @@ StepInfo
    :members:
 
 .. autoclass:: tango.step_info.StepState
+   :member-order: bysource
+   :members:
+
+.. autoclass:: tango.step_info.PlatformMetadata
+   :member-order: bysource
+   :members:
+
+.. autoclass:: tango.step_info.EnvironmentMetadata
+   :member-order: bysource
+   :members:
+
+.. autoclass:: tango.step_info.GitMetadata
+   :member-order: bysource
+   :members:
+
+.. autoclass:: tango.step_info.TangoMetadata
+   :member-order: bysource
+   :members:
