+++ docs/source/api/components/workspace.rst
@@ -19,15 +19,3 @@ Metadata
 
 .. autoclass:: tango.workspace.Run
    :members:
-
-.. autoclass:: tango.workspace.StepExecutionMetadata
-   :members:
-
-.. autoclass:: tango.workspace.TangoMetadata
-   :members:
-
-.. autoclass:: tango.workspace.PlatformMetadata
-   :members:
-
-.. autoclass:: tango.workspace.GitMetadata
-   :members:
