+++ tests/main_test.py
@@ -12,7 +12,6 @@ from tango.common import Params
 from tango.common.testing import TangoTestCase
 from tango.settings import TangoGlobalSettings
 from tango.version import VERSION
-from tango.workspace import StepExecutionMetadata
 
 
 class TestRun(TangoTestCase):
@@ -136,22 +135,6 @@ class TestRun(TangoTestCase):
         assert (run_dir / "hello" / "cache-metadata.json").is_file()
         assert (run_dir / "hello" / "execution-metadata.json").is_file()
         assert (run_dir / "hello_world").is_dir()
-
-        # Check metadata.
-        metadata_path = run_dir / "hello_world" / "execution-metadata.json"
-        assert metadata_path.is_file()
-        metadata_params = Params.from_file(metadata_path)
-        metadata = StepExecutionMetadata.from_params(metadata_params)
-        assert metadata.config == {
-            "type": "concat_strings",
-            "step_unique_id_override": "ConcatStringsStep-3qLNS3gLUQXBGN7rC8yzuYJnjTVp7kgu",
-            "string1": {"type": "ref", "ref": "StringStep-4cHbmoHigd3rvNn3w7shc1d45WA1ijSp"},
-            "string2": "World!",
-            "join_with": ", ",
-        }
-        if (Path.cwd() / ".git").exists():
-            assert metadata.git.commit is not None
-            assert metadata.git.remote is not None
 
         # Check for requirements.txt file.
         assert (run_dir / "hello_world" / "requirements.txt").is_file()
